requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
python-decouple
dj-database-url
django-cors-headers
djangorestframework-simplejwt